+++ pub/src/main/java/com/grupo_5/pub/Repository/ComandaRepository.java
@@ -3,6 +3,6 @@ package com.grupo_5.pub.Repository;
 import com.grupo_5.pub.Model.Comanda;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComandaRepository extends JpaRepository<Comanda, Long> {
+public interface ComandaRepository extends JpaRepository<Comanda, Integer> {
 
 }
