+++ pub/src/main/java/com/grupo_5/pub/Repository/ItemComandaRepository.java
@@ -3,5 +3,5 @@ package com.grupo_5.pub.Repository;
 import com.grupo_5.pub.Model.ItemComanda;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemComandaRepository extends JpaRepository<ItemComanda, Long> {
+public interface ItemComandaRepository extends JpaRepository<ItemComanda, Integer> {
 }
